truc

+++ Tronc Commun/Rank 0/ft_libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: basverdi <basverdi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 00:52:05 by basverdi          #+#    #+#             */
-/*   Updated: 2024/03/01 15:58:27 by basverdi         ###   ########.fr       */
+/*   Updated: 2024/03/11 17:48:13 by basverdi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,13 +63,13 @@ char		*ft_strdup(const char *s);
 char		*ft_strjoin(const char *s1, const char *s2);
 size_t		ft_strlcat(char *dest, const char *src, size_t size);
 size_t		ft_strlcpy(char *dest, const char *src, size_t size);
-size_t		ft_strlen(const char *str);
+int			ft_strlen(const char *str);
 char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 int			ft_strncmp(const char *s1, const char *s2, size_t n);
 char		*ft_strnstr(const char *s1, const char *s2, size_t len);
 char		*ft_strrchr(const char *s, int c);
 char		*ft_strtrim(const char *s1, const char *set);
-char		*ft_substr(const char *s, unsigned int start, size_t len);
+char		*ft_substr(const char *s, int start, int len);
 int			ft_tolower(int c);
 int			ft_toupper(int c);
