+++ Tronc Commun/Rank 0/ft_libft/string/ft_substr.c
@@ -6,16 +6,16 @@
 /*   By: basverdi <basverdi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 01:16:26 by bastienverd       #+#    #+#             */
-/*   Updated: 2024/03/01 15:59:21 by basverdi         ###   ########.fr       */
+/*   Updated: 2024/03/11 17:48:00 by basverdi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-char	*ft_substr(const char *s, unsigned int start, size_t len)
+char	*ft_substr(const char *s, int start, int len)
 {
 	char	*dest;
-	size_t	i;
+	int		i;
 
 	i = 0;
 	if (!s)
